feat(feature-manager): remove Captcha from feature manager (#2157)
* feat(feature-manager): remove Captcha from feature manager

* build: add test credentials for captcha

* refactor: move Captcha Joi validation into controllers
