+++ src/app/config/feature-manager/captcha.config.ts
@@ -1,21 +1,25 @@
-import { FeatureNames, RegisterableFeature } from './types'
+import convict, { Schema } from 'convict'
 
-const captchaFeature: RegisterableFeature<FeatureNames.Captcha> = {
-  name: FeatureNames.Captcha,
-  schema: {
-    captchaPrivateKey: {
-      doc: 'Google Captcha private key',
-      format: String,
-      default: null,
-      env: 'GOOGLE_CAPTCHA',
-    },
-    captchaPublicKey: {
-      doc: 'Google Captcha public key.',
-      format: String,
-      default: '6Lchkl0UAAAAANw4DUdja4W6A5xFwuomkcaiqnkC',
-      env: 'GOOGLE_CAPTCHA_PUBLIC',
-    },
+export interface ICaptcha {
+  captchaPrivateKey: string
+  captchaPublicKey: string
+}
+
+const captchaSchema: Schema<ICaptcha> = {
+  captchaPrivateKey: {
+    doc: 'Google Captcha private key',
+    format: String,
+    default: null,
+    env: 'GOOGLE_CAPTCHA',
+  },
+  captchaPublicKey: {
+    doc: 'Google Captcha public key.',
+    format: String,
+    default: '6Lchkl0UAAAAANw4DUdja4W6A5xFwuomkcaiqnkC',
+    env: 'GOOGLE_CAPTCHA_PUBLIC',
   },
 }
 
-export default captchaFeature
+export const captchaConfig = convict(captchaSchema)
+  .validate({ allowed: 'strict' })
+  .getProperties()
