+++ src/app/config/feature-manager/index.ts
@@ -1,5 +1,4 @@
 import FeatureManager from './util/FeatureManager.class'
-import captcha from './captcha.config'
 import sms from './sms.config'
 import spcpMyInfo from './spcp-myinfo.config'
 import verifiedFields from './verified-fields.config'
@@ -10,7 +9,6 @@ export * from './types'
 const featureManager = new FeatureManager()
 
 // Register features and associated middleware/fallbacks
-featureManager.register(captcha)
 featureManager.register(spcpMyInfo)
 featureManager.register(webhookVerifiedContent)
 featureManager.register(sms)
