+++ src/app/modules/submission/email-submission/email-submission.routes.ts
@@ -1,7 +1,6 @@
 import { Router } from 'express'
 
 import { rateLimitConfig } from '../../../config/config'
-import { CaptchaFactory } from '../../../services/captcha/captcha.factory'
 import { limitRate } from '../../../utils/limit-rate'
 
 import { handleEmailSubmission } from './email-submission.controller'
@@ -29,6 +28,5 @@ export const EmailSubmissionRouter = Router()
 EmailSubmissionRouter.post(
   '/:formId([a-fA-F0-9]{24})',
   limitRate({ max: rateLimitConfig.submissions }),
-  CaptchaFactory.validateCaptchaParams,
   handleEmailSubmission,
 )
