+++ src/app/modules/submission/encrypt-submission/encrypt-submission.routes.ts
@@ -1,7 +1,6 @@
 import { Router } from 'express'
 
 import { rateLimitConfig } from '../../../config/config'
-import { CaptchaFactory } from '../../../services/captcha/captcha.factory'
 import { limitRate } from '../../../utils/limit-rate'
 
 import * as EncryptSubmissionController from './encrypt-submission.controller'
@@ -24,6 +23,5 @@ export const EncryptSubmissionRouter = Router()
 EncryptSubmissionRouter.post(
   '/:formId([a-fA-F0-9]{24})',
   limitRate({ max: rateLimitConfig.submissions }),
-  CaptchaFactory.validateCaptchaParams,
   EncryptSubmissionController.handleEncryptedSubmission,
 )
