+++ src/app/services/captcha/captcha.middleware.ts
@@ -1,0 +1,7 @@
+import { celebrate, Joi, Segments } from 'celebrate'
+
+export const validateCaptchaParams = celebrate({
+  [Segments.QUERY]: Joi.object({
+    captchaResponse: Joi.string().allow(null).required(),
+  }),
+})
